pibot-voice/requirements.txt
# Pibot Voice - Python Dependencies

# Core
pyyaml>=6.0
aiohttp>=3.8.0

# Text-to-Speech
edge-tts>=6.1.0

# Speech-to-Text (choose one)
# Local Whisper (recommended for Pi 5)
openai-whisper>=20231117
# OR API mode
openai>=1.0.0

# Audio I/O
pyaudio>=0.2.13

# Wake Word (optional)
# openwakeword>=0.5.0  # Uncomment if using OpenWakeWord

# Development
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
